requirements.txt
# 雌小鬼的电梯按钮检测系统依赖文件 🎀
# 安装命令: pip install -r requirements.txt

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# 计算机视觉和图像处理
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Web界面
streamlit>=1.28.0

# 数据处理和工具
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0

# 可选：GPU加速（如果有NVIDIA GPU）
# torch-audio  # 如果需要音频处理
# torchtext    # 如果需要文本处理

# 开发工具（可选）
# pytest>=7.0.0          # 测试框架
# black>=23.0.0           # 代码格式化
# ruff>=0.1.0             # 代码检查
# mypy>=1.0.0             # 类型检查

